Extract win32 localatiztion to a new file win32locale.py.
