+++ gramps/gen/utils/grampstranslation.py
@@ -2,9 +2,7 @@
 #
 # Gramps - a GTK+/GNOME based genealogy program
 #
-# Copyright (C) 2000-2006  Donald N. Allingham
-# Copyright (C) 2009       Brian G. Matherly
-# Copyright (C) 2013       John Ralls
+# Copyright (C) 2023       John Ralls
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
